functions/requirements.txt
blinker==1.7.0
boto3==1.34.44
botocore==1.34.44
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
cryptography==42.0.3
deprecation==2.1.0
firebase-admin==6.4.0
firebase-functions==0.1.2
Flask==3.0.2
Flask-Cors==4.0.0
functions-framework==3.5.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
msgpack==1.0.7
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
watchdog==4.0.0
Werkzeug==3.0.1